+++ lottery/src/test/groovy/com/example/lottery/app/LotteryControllerTest.groovy
@@ -1,4 +1,4 @@
-package com.example.lottery
+package com.example.lottery.app
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
